Change how create-topics.sh checks for running Kafka
It seems in Kafka 0.8, the port is bound early in the process of
starting up the server, so sometimes create-topics.sh will proceed to
start creating the topics before the Kafka server is up, which causes
errors. If the broker ID is set, then a perhaps more reliable way to
check that the server is up and running is to check that the ephemeral
ZooKeeper node /brokers/ids/<ID> exists. This is the node that Kafka
uses to check the health of other brokers in the cluster, so if the
node is there, then that indicates the broker is up and healthy.
(If the broker ID is not set, then unfortunately, there doesn't
appear to be a reliable way to check which broker ID was assigned, so
checking for port binding is a reasonable fallback.)