Allow cross building kafka for various archs
This commit introduces a way by which we can now build multi-arch
docker images of kakfa. This patch provides an initial framework
based on some inspiration from:
https://github.com/kubernetes/kubernetes/tree/master/build/debian-base

This patch will allow you to build a cross platform image on an
amd64 machine.

Usage:

     `make build` ==> Retains the old behavior of builds
     `make build ARCH=ppc64le` ==> Builds it for a specific arch

We can further include s390x and arm as other platforms.
Further we can use the manifest-tool or the latest docker client
to push manifest images to dockerhub.