Create change
GitHub Pull Request: https://api.github.com/repos/wurstmeister/kafka-docker/pulls/160

Change how create-topics.sh checks for running Kafka

It seems in Kafka 0.8, the port is bound early in the process of starting up the server (I saw this happening in `network_mode: host`, which may have something to do with it), so sometimes `create-topics.sh` will proceed to start creating the topics before the Kafka server is up, which causes errors. If the broker ID is set, then a perhaps more reliable way to check that the server is up and running is to check that the ephemeral ZooKeeper node `/brokers/ids/<ID>` exists. This is the node that Kafka uses to check the health of other brokers in the cluster, so if the node is there, then that indicates the broker is up and healthy. If the broker ID is not set, then unfortunately, there doesn't appear to be a reliable way to check which broker ID was assigned, so checking for port binding is a reasonable fallback. Maybe another option in this case is just grepping the server logs for a line that looks like



```

[2017-01-28 20:57:19,354] INFO [Kafka Server 0], started (kafka.server.KafkaServer)

```



which is kind of hacky, but may still be more reliable.

Patch-set: 1
Change-id: If9e4ad4a644cf142a717ca3e4d9a6dff68384b1e
Subject: Change how create-topics.sh checks for running Kafka
Branch: refs/heads/master
Commit: 0022d01a0271b5e587dd9cc9ec02181fbf7634e6
Tag: autogenerated:gerrit:newPatchSet
Groups: 0022d01a0271b5e587dd9cc9ec02181fbf7634e6
Work-in-progress: false